turn off persistence with :memory: path

+++ store_test.go
@@ -10,15 +10,21 @@ import (
 	"github.com/hashicorp/raft"
 )
 
-func testFastLogStore(t testing.TB) *FastLogStore {
-	fh, err := ioutil.TempFile("", "bunt")
-	if err != nil {
-		t.Fatalf("err: %s", err)
+func testFastLogStore(t testing.TB, inmem bool) *FastLogStore {
+	var name string
+	if inmem {
+		name = ":memory:"
+	} else {
+		fh, err := ioutil.TempFile("", "bunt")
+		if err != nil {
+			t.Fatalf("err: %s", err)
+		}
+		os.Remove(fh.Name())
+		name = fh.Name()
 	}
-	os.Remove(fh.Name())
 
 	// Successfully creates and returns a store
-	store, err := NewFastLogStore(fh.Name(), Medium, nil)
+	store, err := NewFastLogStore(name, Medium, nil)
 	if err != nil {
 		t.Fatalf("err: %s", err)
 	}
@@ -72,315 +78,318 @@ func TestNewFastLogStore(t *testing.T) {
 	if err := store.Close(); err != nil {
 		t.Fatalf("err: %s", err)
 	}
-
-	//// Ensure our tables were created
-	//db, err := bunt.Open(fh.Name(), dbFileMode, nil)
-	//if err != nil {
-	//	t.Fatalf("err: %s", err)
-	//}
-	//tx, err := db.Begin(true)
-	//if err != nil {
-	//	t.Fatalf("err: %s", err)
-	//}
-	//if _, err := tx.CreateBucket([]byte(dbLogs)); err != bunt.ErrBucketExists {
-	//	t.Fatalf("bad: %v", err)
-	//}
-	//if _, err := tx.CreateBucket([]byte(dbConf)); err != bunt.ErrBucketExists {
-	//	t.Fatalf("bad: %v", err)
-	//}
 }
 
 func TestFastLogStore_Peers(t *testing.T) {
-	store := testFastLogStore(t)
-	defer store.Close()
-	defer os.Remove(store.path)
-	peers, err := store.Peers()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(peers) != 0 {
-		t.Fatalf("expected '%v', got '%v'", 0, len(peers))
-	}
-	v := []string{"1", "2", "3"}
-	if err := store.SetPeers(v); err != nil {
-		t.Fatal(err)
-	}
-	peers, err = store.Peers()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(peers) != 3 {
-		t.Fatalf("expected '%v', got '%v'", 3, len(peers))
-	}
-	if peers[0] != "1" || peers[1] != "2" || peers[2] != "3" {
-		t.Fatalf("expected %v, got %v", v, peers)
+	for p := 0; p < 2; p++ {
+		store := testFastLogStore(t, p == 1)
+		defer store.Close()
+		defer os.Remove(store.path)
+		peers, err := store.Peers()
+		if err != nil {
+			t.Fatal(err)
+		}
+		if len(peers) != 0 {
+			t.Fatalf("expected '%v', got '%v'", 0, len(peers))
+		}
+		v := []string{"1", "2", "3"}
+		if err := store.SetPeers(v); err != nil {
+			t.Fatal(err)
+		}
+		peers, err = store.Peers()
+		if err != nil {
+			t.Fatal(err)
+		}
+		if len(peers) != 3 {
+			t.Fatalf("expected '%v', got '%v'", 3, len(peers))
+		}
+		if peers[0] != "1" || peers[1] != "2" || peers[2] != "3" {
+			t.Fatalf("expected %v, got %v", v, peers)
+		}
 	}
 }
 
 func TestFastLogStore_Persist(t *testing.T) {
-	store := testFastLogStore(t)
-	defer store.Close()
-	defer os.Remove(store.path)
+	for p := 0; p < 2; p++ {
+		store := testFastLogStore(t, p == 1)
+		defer store.Close()
+		defer os.Remove(store.path)
 
-	// Set a mock raft log
-	logs := []*raft.Log{
-		testRaftLog(1, "log1"),
-		testRaftLog(2, "log2"),
-		testRaftLog(3, "log3"),
-		testRaftLog(4, "log4"),
+		// Set a mock raft log
+		logs := []*raft.Log{
+			testRaftLog(1, "log1"),
+			testRaftLog(2, "log2"),
+			testRaftLog(3, "log3"),
+			testRaftLog(4, "log4"),
+		}
+
+		if err := store.StoreLogs(logs); err != nil {
+			t.Fatalf("bad: %s", err)
+		}
+
+		store.Close()
+		var err error
+		store, err = NewFastLogStore(store.path, Medium, nil)
+		if err != nil {
+			t.Fatal(err)
+		}
 	}
-
-	if err := store.StoreLogs(logs); err != nil {
-		t.Fatalf("bad: %s", err)
-	}
-
-	store.Close()
-	var err error
-	store, err = NewFastLogStore(store.path, Medium, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-
 }
 func TestFastLogStore_FirstIndex(t *testing.T) {
-	store := testFastLogStore(t)
-	defer store.Close()
-	defer os.Remove(store.path)
+	for p := 0; p < 2; p++ {
+		store := testFastLogStore(t, p == 1)
+		defer store.Close()
+		defer os.Remove(store.path)
 
-	// Should get 0 index on empty log
-	idx, err := store.FirstIndex()
-	if err != nil {
-		t.Fatalf("err: %s", err)
-	}
-	if idx != 0 {
-		t.Fatalf("bad: %v", idx)
-	}
+		// Should get 0 index on empty log
+		idx, err := store.FirstIndex()
+		if err != nil {
+			t.Fatalf("err: %s", err)
+		}
+		if idx != 0 {
+			t.Fatalf("bad: %v", idx)
+		}
 
-	// Set a mock raft log
-	logs := []*raft.Log{
-		testRaftLog(1, "log1"),
-		testRaftLog(2, "log2"),
-		testRaftLog(3, "log3"),
-	}
-	if err := store.StoreLogs(logs); err != nil {
-		t.Fatalf("bad: %s", err)
-	}
+		// Set a mock raft log
+		logs := []*raft.Log{
+			testRaftLog(1, "log1"),
+			testRaftLog(2, "log2"),
+			testRaftLog(3, "log3"),
+		}
+		if err := store.StoreLogs(logs); err != nil {
+			t.Fatalf("bad: %s", err)
+		}
 
-	// Fetch the first Raft index
-	idx, err = store.FirstIndex()
-	if err != nil {
-		t.Fatalf("err: %s", err)
-	}
-	if idx != 1 {
-		t.Fatalf("bad: %d", idx)
+		// Fetch the first Raft index
+		idx, err = store.FirstIndex()
+		if err != nil {
+			t.Fatalf("err: %s", err)
+		}
+		if idx != 1 {
+			t.Fatalf("bad: %d", idx)
+		}
 	}
 }
 
 func TestFastLogStore_LastIndex(t *testing.T) {
-	store := testFastLogStore(t)
-	defer store.Close()
-	defer os.Remove(store.path)
+	for p := 0; p < 2; p++ {
+		store := testFastLogStore(t, p == 1)
+		defer store.Close()
+		defer os.Remove(store.path)
 
-	// Should get 0 index on empty log
-	idx, err := store.LastIndex()
-	if err != nil {
-		t.Fatalf("err: %s", err)
-	}
-	if idx != 0 {
-		t.Fatalf("bad: %v", idx)
-	}
+		// Should get 0 index on empty log
+		idx, err := store.LastIndex()
+		if err != nil {
+			t.Fatalf("err: %s", err)
+		}
+		if idx != 0 {
+			t.Fatalf("bad: %v", idx)
+		}
 
-	// Set a mock raft log
-	logs := []*raft.Log{
-		testRaftLog(1, "log1"),
-		testRaftLog(2, "log2"),
-		testRaftLog(3, "log3"),
-	}
-	if err := store.StoreLogs(logs); err != nil {
-		t.Fatalf("bad: %s", err)
-	}
+		// Set a mock raft log
+		logs := []*raft.Log{
+			testRaftLog(1, "log1"),
+			testRaftLog(2, "log2"),
+			testRaftLog(3, "log3"),
+		}
+		if err := store.StoreLogs(logs); err != nil {
+			t.Fatalf("bad: %s", err)
+		}
 
-	// Fetch the last Raft index
-	idx, err = store.LastIndex()
-	if err != nil {
-		t.Fatalf("err: %s", err)
-	}
-	if idx != 3 {
-		t.Fatalf("bad: %d", idx)
+		// Fetch the last Raft index
+		idx, err = store.LastIndex()
+		if err != nil {
+			t.Fatalf("err: %s", err)
+		}
+		if idx != 3 {
+			t.Fatalf("bad: %d", idx)
+		}
 	}
 }
 
 func TestFastLogStore_GetLog(t *testing.T) {
-	store := testFastLogStore(t)
-	defer store.Close()
-	defer os.Remove(store.path)
+	for p := 0; p < 2; p++ {
+		store := testFastLogStore(t, p == 1)
+		defer store.Close()
+		defer os.Remove(store.path)
 
-	log := new(raft.Log)
+		log := new(raft.Log)
 
-	// Should return an error on non-existent log
-	if err := store.GetLog(1, log); err != raft.ErrLogNotFound {
-		t.Fatalf("expected raft log not found error, got: %v", err)
-	}
+		// Should return an error on non-existent log
+		if err := store.GetLog(1, log); err != raft.ErrLogNotFound {
+			t.Fatalf("expected raft log not found error, got: %v", err)
+		}
 
-	// Set a mock raft log
-	logs := []*raft.Log{
-		testRaftLog(1, "log1"),
-		testRaftLog(2, "log2"),
-		testRaftLog(3, "log3"),
-	}
-	if err := store.StoreLogs(logs); err != nil {
-		t.Fatalf("bad: %s", err)
-	}
+		// Set a mock raft log
+		logs := []*raft.Log{
+			testRaftLog(1, "log1"),
+			testRaftLog(2, "log2"),
+			testRaftLog(3, "log3"),
+		}
+		if err := store.StoreLogs(logs); err != nil {
+			t.Fatalf("bad: %s", err)
+		}
 
-	// Should return the proper log
-	if err := store.GetLog(2, log); err != nil {
-		t.Fatalf("err: %s", err)
-	}
-	if !reflect.DeepEqual(log, logs[1]) {
-		t.Fatalf("bad: %#v\n%#v", log, logs[1])
+		// Should return the proper log
+		if err := store.GetLog(2, log); err != nil {
+			t.Fatalf("err: %s", err)
+		}
+		if !reflect.DeepEqual(log, logs[1]) {
+			t.Fatalf("bad: %#v\n%#v", log, logs[1])
+		}
 	}
 }
 
 func TestFastLogStore_SetLog(t *testing.T) {
-	store := testFastLogStore(t)
-	defer store.Close()
-	defer os.Remove(store.path)
+	for p := 0; p < 2; p++ {
+		store := testFastLogStore(t, p == 1)
+		defer store.Close()
+		defer os.Remove(store.path)
 
-	// Create the log
-	log := &raft.Log{
-		Data:  []byte("log1"),
-		Index: 1,
-	}
+		// Create the log
+		log := &raft.Log{
+			Data:  []byte("log1"),
+			Index: 1,
+		}
 
-	// Attempt to store the log
-	if err := store.StoreLog(log); err != nil {
-		t.Fatalf("err: %s", err)
-	}
+		// Attempt to store the log
+		if err := store.StoreLog(log); err != nil {
+			t.Fatalf("err: %s", err)
+		}
 
-	// Retrieve the log again
-	result := new(raft.Log)
-	if err := store.GetLog(1, result); err != nil {
-		t.Fatalf("err: %s", err)
-	}
+		// Retrieve the log again
+		result := new(raft.Log)
+		if err := store.GetLog(1, result); err != nil {
+			t.Fatalf("err: %s", err)
+		}
 
-	// Ensure the log comes back the same
-	if !reflect.DeepEqual(log, result) {
-		t.Fatalf("bad: %v", result)
+		// Ensure the log comes back the same
+		if !reflect.DeepEqual(log, result) {
+			t.Fatalf("bad: %v", result)
+		}
 	}
 }
 
 func TestFastLogStore_SetLogs(t *testing.T) {
-	store := testFastLogStore(t)
-	defer store.Close()
-	defer os.Remove(store.path)
+	for p := 0; p < 2; p++ {
+		store := testFastLogStore(t, p == 1)
+		defer store.Close()
+		defer os.Remove(store.path)
 
-	// Create a set of logs
-	logs := []*raft.Log{
-		testRaftLog(1, "log1"),
-		testRaftLog(2, "log2"),
-	}
+		// Create a set of logs
+		logs := []*raft.Log{
+			testRaftLog(1, "log1"),
+			testRaftLog(2, "log2"),
+		}
 
-	// Attempt to store the logs
-	if err := store.StoreLogs(logs); err != nil {
-		t.Fatalf("err: %s", err)
-	}
+		// Attempt to store the logs
+		if err := store.StoreLogs(logs); err != nil {
+			t.Fatalf("err: %s", err)
+		}
 
-	// Ensure we stored them all
-	result1, result2 := new(raft.Log), new(raft.Log)
-	if err := store.GetLog(1, result1); err != nil {
-		t.Fatalf("err: %s", err)
-	}
-	if !reflect.DeepEqual(logs[0], result1) {
-		t.Fatalf("bad: %#v", result1)
-	}
-	if err := store.GetLog(2, result2); err != nil {
-		t.Fatalf("err: %s", err)
-	}
-	if !reflect.DeepEqual(logs[1], result2) {
-		t.Fatalf("bad: %#v", result2)
+		// Ensure we stored them all
+		result1, result2 := new(raft.Log), new(raft.Log)
+		if err := store.GetLog(1, result1); err != nil {
+			t.Fatalf("err: %s", err)
+		}
+		if !reflect.DeepEqual(logs[0], result1) {
+			t.Fatalf("bad: %#v", result1)
+		}
+		if err := store.GetLog(2, result2); err != nil {
+			t.Fatalf("err: %s", err)
+		}
+		if !reflect.DeepEqual(logs[1], result2) {
+			t.Fatalf("bad: %#v", result2)
+		}
 	}
 }
 
 func TestFastLogStore_DeleteRange(t *testing.T) {
-	store := testFastLogStore(t)
-	defer store.Close()
-	defer os.Remove(store.path)
+	for p := 0; p < 2; p++ {
+		store := testFastLogStore(t, p == 1)
+		defer store.Close()
+		defer os.Remove(store.path)
 
-	// Create a set of logs
-	log1 := testRaftLog(1, "log1")
-	log2 := testRaftLog(2, "log2")
-	log3 := testRaftLog(3, "log3")
-	logs := []*raft.Log{log1, log2, log3}
+		// Create a set of logs
+		log1 := testRaftLog(1, "log1")
+		log2 := testRaftLog(2, "log2")
+		log3 := testRaftLog(3, "log3")
+		logs := []*raft.Log{log1, log2, log3}
 
-	// Attempt to store the logs
-	if err := store.StoreLogs(logs); err != nil {
-		t.Fatalf("err: %s", err)
-	}
+		// Attempt to store the logs
+		if err := store.StoreLogs(logs); err != nil {
+			t.Fatalf("err: %s", err)
+		}
 
-	// Attempt to delete a range of logs
-	if err := store.DeleteRange(1, 2); err != nil {
-		t.Fatalf("err: %s", err)
-	}
+		// Attempt to delete a range of logs
+		if err := store.DeleteRange(1, 2); err != nil {
+			t.Fatalf("err: %s", err)
+		}
 
-	// Ensure the logs were deleted
-	if err := store.GetLog(1, new(raft.Log)); err != raft.ErrLogNotFound {
-		t.Fatalf("should have deleted log1")
-	}
-	if err := store.GetLog(2, new(raft.Log)); err != raft.ErrLogNotFound {
-		t.Fatalf("should have deleted log2")
+		// Ensure the logs were deleted
+		if err := store.GetLog(1, new(raft.Log)); err != raft.ErrLogNotFound {
+			t.Fatalf("should have deleted log1")
+		}
+		if err := store.GetLog(2, new(raft.Log)); err != raft.ErrLogNotFound {
+			t.Fatalf("should have deleted log2")
+		}
 	}
 }
 
 func TestFastLogStore_Set_Get(t *testing.T) {
-	store := testFastLogStore(t)
-	defer store.Close()
-	defer os.Remove(store.path)
+	for p := 0; p < 2; p++ {
+		store := testFastLogStore(t, p == 1)
+		defer store.Close()
+		defer os.Remove(store.path)
 
-	// Returns error on non-existent key
-	if _, err := store.Get([]byte("bad")); err != ErrKeyNotFound {
-		t.Fatalf("expected not found error, got: %q", err)
-	}
+		// Returns error on non-existent key
+		if _, err := store.Get([]byte("bad")); err != ErrKeyNotFound {
+			t.Fatalf("expected not found error, got: %q", err)
+		}
 
-	k, v := []byte("hello"), []byte("world")
+		k, v := []byte("hello"), []byte("world")
 
-	// Try to set a k/v pair
-	if err := store.Set(k, v); err != nil {
-		t.Fatalf("err: %s", err)
-	}
+		// Try to set a k/v pair
+		if err := store.Set(k, v); err != nil {
+			t.Fatalf("err: %s", err)
+		}
 
-	// Try to read it back
-	val, err := store.Get(k)
-	if err != nil {
-		t.Fatalf("err: %s", err)
-	}
-	if !bytes.Equal(val, v) {
-		t.Fatalf("bad: %v", val)
+		// Try to read it back
+		val, err := store.Get(k)
+		if err != nil {
+			t.Fatalf("err: %s", err)
+		}
+		if !bytes.Equal(val, v) {
+			t.Fatalf("bad: %v", val)
+		}
 	}
 }
 
 func TestFastLogStore_SetUint64_GetUint64(t *testing.T) {
-	store := testFastLogStore(t)
-	defer store.Close()
-	defer os.Remove(store.path)
+	for p := 0; p < 2; p++ {
+		store := testFastLogStore(t, p == 1)
+		defer store.Close()
+		defer os.Remove(store.path)
 
-	// Returns error on non-existent key
-	if _, err := store.GetUint64([]byte("bad")); err != ErrKeyNotFound {
-		t.Fatalf("expected not found error, got: %q", err)
-	}
+		// Returns error on non-existent key
+		if _, err := store.GetUint64([]byte("bad")); err != ErrKeyNotFound {
+			t.Fatalf("expected not found error, got: %q", err)
+		}
 
-	k, v := []byte("abc"), uint64(123)
+		k, v := []byte("abc"), uint64(123)
 
-	// Attempt to set the k/v pair
-	if err := store.SetUint64(k, v); err != nil {
-		t.Fatalf("err: %s", err)
-	}
+		// Attempt to set the k/v pair
+		if err := store.SetUint64(k, v); err != nil {
+			t.Fatalf("err: %s", err)
+		}
 
-	// Read back the value
-	val, err := store.GetUint64(k)
-	if err != nil {
-		t.Fatalf("err: %s", err)
-	}
-	if val != v {
-		t.Fatalf("bad: %v", val)
+		// Read back the value
+		val, err := store.GetUint64(k)
+		if err != nil {
+			t.Fatalf("err: %s", err)
+		}
+		if val != v {
+			t.Fatalf("bad: %v", val)
+		}
 	}
 }
